requirements.txt
Django==2.2.9   
djangorestframework==3.11.0
djangorestframework-gis==0.15
psycopg2-binary==2.8.3
pytz==2019.3
sqlparse==0.3.0
django-filter==2.2.0
djoser==2.0.3
djangorestframework_simplejwt==4.4.0
Pillow==2.1.0
django-imagekit==4.0.2
django_filters==0.2.1
django-filter==2.2.0
django-cors-headers==3.0.1
